requirements.txt
Pillow
fastapi
uvicorn
tensorflow
pydantic
opencv-python
numpy
pandas
scikit-learn
opencv-python
numpy
scikit-learn
reportlab
passlib[bcrypt]
pyjwt
SQLAlchemy
mysqlclient
python-dotenv
imageio